move more logic into api service
