+++ core/src/main/java/com/google/cloud/sql/core/SqlAdminApiFetcher.java
@@ -26,21 +26,44 @@ import com.google.auth.oauth2.OAuth2Credentials;
 import com.google.cloud.sql.AuthType;
 import com.google.common.base.CharMatcher;
 import com.google.common.io.BaseEncoding;
+import com.google.common.util.concurrent.FutureCallback;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.ListeningScheduledExecutorService;
+import com.google.common.util.concurrent.SettableFuture;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.security.GeneralSecurityException;
 import java.security.KeyPair;
+import java.security.KeyStore;
+import java.security.KeyStore.PasswordProtection;
+import java.security.KeyStore.PrivateKeyEntry;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.Callable;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.net.ssl.KeyManagerFactory;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManagerFactory;
+import org.checkerframework.checker.nullness.compatqual.NullableDecl;
 
 /**
  * Class that encapsulates all logic for interacting with SQLAdmin API.
  */
 public class SqlAdminApiFetcher {
 
+  private static final Logger logger = Logger.getLogger(SqlAdminApiFetcher.class.getName());
   private final SQLAdmin apiClient;
 
   public SqlAdminApiFetcher(SQLAdmin apiClient) {
@@ -69,6 +92,95 @@ public class SqlAdminApiFetcher {
         + BaseEncoding.base64().withSeparator("\n", 64).encode(keyPair.getPublic().getEncoded())
         + "\n"
         + "-----END RSA PUBLIC KEY-----\n";
+  }
+
+  // Schedules task to be executed once the provided futures are complete.
+  private <T> ListenableFuture<T> whenAllSucceed(
+      Callable<T> task,
+      ListeningScheduledExecutorService executor,
+      ListenableFuture<?>... futures) {
+    SettableFuture<T> taskFuture = SettableFuture.create();
+
+    // Create a countDown for all Futures to complete.
+    AtomicInteger countDown = new AtomicInteger(futures.length);
+
+    // Trigger the task when all futures are complete.
+    FutureCallback<Object> runWhenInputAreComplete =
+        new FutureCallback<Object>() {
+          @Override
+          public void onSuccess(@NullableDecl Object o) {
+            if (countDown.decrementAndGet() == 0) {
+              taskFuture.setFuture(executor.submit(task));
+            }
+          }
+
+          @Override
+          public void onFailure(Throwable throwable) {
+            if (!taskFuture.setException(throwable)) {
+              String msg = "Got more than one input failure. Logging failures after the first";
+              logger.log(Level.SEVERE, msg, throwable);
+            }
+          }
+        };
+    for (ListenableFuture<?> future : futures) {
+      Futures.addCallback(future, runWhenInputAreComplete, executor);
+    }
+
+    return taskFuture;
+  }
+
+  ListenableFuture<InstanceData> getInstanceData(
+      CloudSqlInstanceName instanceName,
+      OAuth2Credentials credentials,
+      AuthType authType,
+      ListeningScheduledExecutorService executor,
+      ListenableFuture<KeyPair> keyPair) {
+    ListenableFuture<Metadata> metadataFuture = executor.submit(
+        () -> fetchMetadata(instanceName, authType));
+    ListenableFuture<Certificate> ephemeralCertificateFuture =
+        whenAllSucceed(
+            () -> fetchEphemeralCertificate(Futures.getDone(keyPair), instanceName,
+                credentials, authType), executor, keyPair);
+    // Once the API calls are complete, construct the SSLContext for the sockets
+    ListenableFuture<SslData> sslContextFuture =
+        whenAllSucceed(
+            () ->
+                createSslData(
+                    Futures.getDone(keyPair),
+                    Futures.getDone(metadataFuture),
+                    Futures.getDone(ephemeralCertificateFuture),
+                    instanceName,
+                    authType),
+            executor,
+            keyPair,
+            metadataFuture,
+            ephemeralCertificateFuture);
+    // Once both the SSLContext and Metadata are complete, return the results
+    return whenAllSucceed(
+        () -> {
+
+          // Get expiration value for new cert
+          Certificate ephemeralCertificate = Futures.getDone(ephemeralCertificateFuture);
+          X509Certificate x509Certificate = (X509Certificate) ephemeralCertificate;
+          Date expiration = x509Certificate.getNotAfter();
+
+          if (authType == AuthType.IAM) {
+            expiration = getTokenExpirationTime(credentials)
+                .filter(tokenExpiration -> x509Certificate.getNotAfter().after(tokenExpiration))
+                .orElse(x509Certificate.getNotAfter());
+          }
+
+          return new InstanceData(
+              Futures.getDone(metadataFuture), Futures.getDone(sslContextFuture),
+              expiration);
+        },
+        executor,
+        metadataFuture,
+        sslContextFuture);
+  }
+
+  private Optional<Date> getTokenExpirationTime(OAuth2Credentials credentials) {
+    return Optional.ofNullable(credentials.getAccessToken().getExpirationTime());
   }
 
   public String getApplicationName() {
@@ -172,6 +284,61 @@ public class SqlAdminApiFetcher {
     }
 
     return ephemeralCertificate;
+  }
+
+  /**
+   * Creates a new SslData based on the provided parameters. It contains a SSLContext that will be
+   * used to provide new SSLSockets authorized to connect to a Cloud SQL instance. It also contains
+   * a KeyManagerFactory and a TrustManagerFactory that can be used by drivers to establish an SSL
+   * tunnel.
+   */
+  SslData createSslData(
+      KeyPair keyPair, Metadata metadata, Certificate ephemeralCertificate,
+      CloudSqlInstanceName instanceName, AuthType authType) {
+    try {
+      KeyStore authKeyStore = KeyStore.getInstance(KeyStore.getDefaultType());
+      authKeyStore.load(null, null);
+      KeyStore.PrivateKeyEntry privateKey =
+          new PrivateKeyEntry(keyPair.getPrivate(), new Certificate[]{ephemeralCertificate});
+      authKeyStore.setEntry("ephemeral", privateKey, new PasswordProtection(new char[0]));
+      KeyManagerFactory kmf =
+          KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+      kmf.init(authKeyStore, new char[0]);
+
+      KeyStore trustedKeyStore = KeyStore.getInstance(KeyStore.getDefaultType());
+      trustedKeyStore.load(null, null);
+      trustedKeyStore.setCertificateEntry("instance", metadata.getInstanceCaCertificate());
+      TrustManagerFactory tmf = TrustManagerFactory.getInstance("X.509");
+      tmf.init(trustedKeyStore);
+      SSLContext sslContext;
+
+      try {
+        sslContext = SSLContext.getInstance("TLSv1.3");
+      } catch (NoSuchAlgorithmException ex) {
+        if (authType == AuthType.IAM) {
+          throw new RuntimeException(
+              String.format(
+                  "[%s] Unable to create a SSLContext for the Cloud SQL instance.",
+                  instanceName.getConnectionName())
+                  + " TLSv1.3 is not supported for your Java version and is required to connect"
+                  + " using IAM authentication",
+              ex);
+        } else {
+          logger.warning("TLSv1.3 is not supported for your Java version, fallback to TLSv1.2");
+          sslContext = SSLContext.getInstance("TLSv1.2");
+        }
+      }
+
+      sslContext.init(kmf.getKeyManagers(), tmf.getTrustManagers(), new SecureRandom());
+
+      return new SslData(sslContext, kmf, tmf);
+    } catch (GeneralSecurityException | IOException ex) {
+      throw new RuntimeException(
+          String.format(
+              "[%s] Unable to create a SSLContext for the Cloud SQL instance.",
+              instanceName.getConnectionName()),
+          ex);
+    }
   }
 
   /**
